+++ HISTORY.md
@@ -6,6 +6,7 @@
 * objects: fixed a list-watcher direct-editing offset bug
 * morphic: update the Hand's position on mouse-down - avoid triggering at the origin point
 * symbols: added hooks for dynamic coloring
+* blocks: added blocks-fading support for symbols (under construction)
 
 ### 2020-07-19
 * blocks: blocks-fade-out support for label arrows (under construction)